I thought that mapping a value instance to an enum doesn't work properly, because I got invalid Java code which always used Enum.byId, even if I was already providing two enum instances in the value instance. So I've added some code in InstanceSpec to recognize this case and dont' use Enum.byId.
That was stupid, because the enum keyword is used to map numerical values to enums. If I know I'm already using enum instances, I simply need to not use that keyword and the type detection for the value instance already recognizes the correct enum types even in my "... ? ... : ...".

So I'm reverting the former changes to get inline with the upstream codebase again.